linux

+++ src/mir.c
@@ -1468,7 +1468,7 @@ is_load_needed(MirInstr *instr)
 		 * values. We get s32 vs *s32 type mismatch without this.
 		 *
 		 * Ex.: j : *s32 = ^ (cast(**s32) i_ptr_ptr);
-		 * 
+		 *
 		 * But I'm not 100% sure that this didn't broke something else...
 		 */
 		MirInstrLoad *load = (MirInstrLoad *)instr;
@@ -5517,7 +5517,7 @@ analyze_instr_switch(Context *cnt, MirInstrSwitch *sw)
 	}
 
 	s64 expected_case_count = expected_case_type->kind == MIR_TYPE_ENUM
-	                              ? expected_case_type->data.enm.variants->size
+	                              ? (s64)expected_case_type->data.enm.variants->size
 	                              : -1;
 
 	if ((expected_case_count > (s64)sw->cases->size) && !sw->has_user_defined_default) {
